Move custom result classes to separate file
